Route: - add react-rouer dependencies - Move Blog components under Blog/ - Setup default route
