+++ package.json
@@ -6,6 +6,8 @@
     "axios": "^0.18.0",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "1.0.13"
   },
   "scripts": {
