+++ src/containers/Blog/Blog.css
@@ -1,11 +1,3 @@
-.Posts {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: center;
-    width: 80%;
-    margin: auto;
-}
-
 .Blog ul {
     list-style: none;
     margin: 0;
