+++ src/containers/Blog/Blog.js
@@ -1,56 +1,12 @@
 import React, { Component } from 'react';
-// import axios from 'axios';
-import axios from '../../axios';
+import { Route } from 'react-router-dom';
 
-import Post from '../../components/Post/Post';
-import FullPost from '../../components/FullPost/FullPost';
-import NewPost from '../../components/NewPost/NewPost';
 import './Blog.css';
+import Posts from './Posts/Posts';
+
 
 class Blog extends Component {
-    state = {
-        posts: [],
-        selectedPostId: null,
-        error: false
-    };
-
-    componentDidMount() {
-        axios.get('posts')
-            .then(response => {
-                const orgPosts = response.data.slice(0, 4);
-                const posts = orgPosts.map(post => {
-                    return {
-                        ...post,
-                        author: 'Bill'
-                    };
-                });
-
-                this.setState({ posts, error: false });
-            })
-            .catch(error => {
-                console.log(error);
-                this.setState({ error: true });
-            });
-    }
-
-    postSelectedHandler = (id) => {
-        this.setState({ selectedPostId: id });
-    }
-
     render() {
-        let posts = <p style={{ textAlign: 'center' }}>Something went wrong!</p>;
-
-        if (!this.state.error) {
-            posts = this.state.posts.map(post => {
-                return <Post
-                    key={post.id}
-                    title={post.title}
-                    author={post.author}
-                    clicked={() => this.postSelectedHandler(post.id)}
-                />
-            });
-        }
-
         return (
             <div className="Blog">
                 <header>
@@ -65,15 +21,8 @@ class Blog extends Component {
                         </ul>
                     </nav>
                 </header>
-                <section className="Posts">
-                    {posts}
-                </section>
-                <section>
-                    <FullPost id={this.state.selectedPostId} />
-                </section>
-                <section>
-                    <NewPost />
-                </section>
+                {/* <Route path="/" exact render={() => <h1>Home</h1>} /> */}
+                <Route path="/" exact component={Posts} />
             </div>
         );
     }
